+++ docs/CHANGELOG.md
@@ -22,6 +22,7 @@ ones are marked like "v1.0.0-fork".
 
 * Fixes [#113](https://github.com/HugoFara/lwt/issues/113): changing a language name was making text-to-speech unusable and may prompt a warning.
 * Since 2.9.0, tests may be unable to run due to a missing key in an AJAX request ("type"). This is fixed.
+* Since 2.9.0, audio position was not saved properly.
 
 ### Deprecated